+++ src/lib/data/queries.ts
@@ -204,6 +204,11 @@ export const getMigrationHistoriesByContractAddress = graphql(`
         timestamp
       }
       remark
+      code {
+        account {
+          address
+        }
+      }
     }
   }
 `);